requirements.txt
fastapi==0.115.0
uvicorn==0.30.1
jinja2==3.1.3
transformers==4.44.2
datasets==3.0.1
evaluate
accelerate
torch==2.3.0
scikit-learn==1.5.0
pandas==2.2.2
numpy==1.26.4
matplotlib==3.9.0
huggingface_hub==0.24.6
pytest==8.2.0
httpx
python-multipart
loguru
prometheus_client
python-dotenv
black
prefect
prefect-docker